Make the NetLogo dependency transitive
This pulls in all necessary jar files to compile and test against.  This
way the extension authors do not have to know the "minimal set" of deps
to include if they want to compile and run tests locally.

Removing these extra jars from the packaging process will be handled
separately.